im dying Co-authored-by: Jessica Xu <jessicaxu0605@users.noreply.github.com>

+++ models/ability/ability.h
@@ -36,7 +36,7 @@ protected:
     // shared validation based on numParams()
     void validateNumParams(const std::vector<std::string> &args) const
     {
-        if (args.size() != numParams())
+        if (args.size() != static_cast<std::size_t>(numParams()))
         {
             throw std::invalid_argument(
                 "Ability '" + name() + "' expected " + std::to_string(numParams()) +
